The test sporadically failed with valgrind trying to match `no client named obuf-client1 found*`
in the log it looks like `obuf-client1` was indeed dropped, so i'm guessing it's because CLIENT LIST was processed first.

https://github.com/redis/redis/actions/runs/4199325173/jobs/7284050432