@oranagra 2023-02-20
btw, i noticed that `client evicted due to client tracking prefixes` takes over 200 seconds with valgrind, maybe someone can look into that.