@enjoy-binbin 2023-02-21
> btw, i noticed that client evicted due to client tracking prefixes takes over 200 seconds with valgrind, maybe someone can look into that.

i do some tests in my local machine, there are two ways that we can speed up the testing (did not dive into the source code):
unstable normal: 3508ms, valgrind: 289503ms -> 290s
```c
    test "client evicted due to client tracking prefixes" {
        r flushdb
        set rr [redis_client]
        
        # Since tracking prefixes list is a small overheatd this test uses a minimal maxmemory-clients config
        set temp_maxmemory_clients 200000
        r config set maxmemory-clients $temp_maxmemory_clients
        
        # Append tracking prefixes until list maxes out maxmemroy clients and causes client eviction
        catch { 
            for {set j 0} {$j < $temp_maxmemory_clients} {incr j} {
                $rr client tracking on prefix [format %012s $j] bcast
            }
        } e
        assert_match {I/O error reading reply} $e
        $rr close

        # Restore config for next tests
        r config set maxmemory-clients $maxmemory_clients
    }
```


1. Reduce maxmemory-clients, so we can reach threshold faster
normal: 1000ms, valgrind: 109597ms -> 110s
```diff
- set temp_maxmemory_clients 200000
+ set temp_maxmemory_clients 20000
```


2. Include multiple prefixes in one command, reduce the number of command calls

normal: 1800ms, valgrind: 174024ms -> 174s
```diff
- $rr client tracking on prefix [format %012s $j] bcast
+ $rr client tracking on prefix [format %012s $j] prefix [format a%011s $j] bcast
```
valgrind: 135742ms -> 136s
```diff
+ $rr client tracking on prefix [format a%011s $j] prefix [format b%011s $j] prefix [format c%011s $j] bcast
```


3. or combine them, this should be much faster

valgrind: 67429ms  -> 67s
```diff
+ set temp_maxmemory_clients 20000
+ $rr client tracking on prefix [format a%011s $j] prefix [format b%011s $j] bcast
```