@oranagra 2023-02-21
my instinct is to combine them, but i'm worried that 20kb too low. i'm afraid that it cause cause the non-valgrind test to be less effective.
maybe we can use larger perfixes? i did try `01024s $j` and it didn't seem to help so i dropped it without understanding why..