@enjoy-binbin 2023-02-21
> maybe we can use larger perfixes? i did try 01024s $j and it didn't seem to help so i dropped it without understanding why..

ok, i see, that's the reason
```c
    /* Add memory overhead of the tracking prefixes, this is an underestimation so we don't need to traverse the entire rax */
    if (c->client_tracking_prefixes)
        mem += c->client_tracking_prefixes->numnodes * (sizeof(raxNode) * sizeof(raxNode*));
```


> my instinct is to combine them, but i'm worried that 20kb too low. i'm afraid that it cause cause the non-valgrind test to be less effective.

yes agree, btw, I saw this while researching the above question

```c
    /* Don't allow a too small maxmemory-clients to avoid cases where we can't communicate
     * at all with the server because of bad configuration */
    if (maxmemory_clients_actual < 1024*128)
        maxmemory_clients_actual = 1024*128;
```