@oranagra 2023-02-21
ok, i didn't recall that.
so let's just combine more prefixes in each command.
regarding documenting that limit, other than tests, i suppose that limit isn't relevant for any actual use case, so i don't think we need to mention it. on the other hand, i see the limit for max-bulk-len **is** documented, so i don't care much either way.